requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
Jinja2==3.1.4
python-multipart==0.0.9
openpyxl==3.1.2
pandas==2.2.3
starlette==0.38.4
itsdangerous==2.1.2

